Tools: Make soaring test a subclass of plane test.
